requirements.txt
PySide6
pyqtgraph
pyzmq
numpy
